changed to typescript
